requirements.txt
Flask~=2.1.2
WTForms~=3.0.1
PyMySQL~=1.0.2
python-dotenv~=0.20.0
pydantic~=1.9.1
email-validator~=1.2.1